Add the 'add' button to the card
Based on the basic function of the table, we should have an add
button to let users to add category or sub-categories to the table.
This commit gets the value from the form object of the AddForm component
and make a request for getting back the categories or sub-categories
based on parentId and name.