Add the new product model and APIs in the database
For developing the product page, we need to create the product model
in the Mongo database for receiving whatever data we want, and then
create new APIs for handling different commands for the front end.

This commit adds image, name, description, price, pCategoryId,
categoryId and detail to the product's schema, and then create the new
route called product for handling CRUD commands.