Organize the header title in Redux
Previously, the header title only can be get  by the getTitle function,
which make the code more complex. The better idea is put all the reusable
and logic states in the redux.
This commit add the setHeaderTitle in the action by action creator, and pass
the state and setHeaderTitle action to reducer by store. After that, the left
nav component set the header title and header component get the header title
by the higher-order function called connect from react-redux.