Organize the user in Redux
Previously, the user object can be get by reqLogin function and
then save to the memory storage. The better idea is move the
login asynchronous function to Redux and get the user object via
store.

This commit add the login asynchronous function in the action and
then dispatch to the reducer. The login component can get the user
object by this.props.login from Redux.