Add the search function in the Product page
Normally, the page should has the search function for users to
select what product they want to look for based on different
conditions.

This commit adds the new function for users to select products by
name or description.
This commit adds the two new APIs for search by name and description,
and then add the logic to get a list of products based on name or
description.