Add the details page and status function
Previously, the product page has the status and details but both
of them are static.

This commit builds up the structure of the details page and get
the data from product by passing the products object to the details
component.

This commit makes the status function works by adding the route of
updating the status based on the logic in the client-side.