Fix the bug of the product page
Previously, the detail page has been added into the
product page. However, the left navigation doesn't
have the selected item when refreshing the page.
This commit changes the logic to make the selected
key point to the 'product' item is the index of the
path is belong to the product item.