Set up the model and route for uploading images
Previously, the add product page has been created.
However, we need the upload image function for allowing
users to upload images to match the product.
This commit sets up the logic in the route controller to
set up the AWS S3 and two different routes for uploading
the images and deleting the images.