Hide the upload and remove button in the detail page
Previously, the add and edit page shows the upload area
with upload, preview and remove buttons. However, the detail
page should only shows the preview button in the picture
wall area.
This commit add the new props called 'showPictureWallModal'
and set the logic to hide the upload and remove buttons in
the detail page of the product.