Add the 404 Error page
Previously, the page always be redirect to home
page when the page cannot be found, which is
not the idea way to match the user experience.
This commit add the 404 error page, which is
used to show the 404 info and provide the button
for user to get back to home page.