Fix the duplicated namings for the routes
Previously, when the user try to refresh the pages
for user, role, product and category, the response
was data from backend side instead of the route of
front end side. However, the duplicated namings
of routes between front-end and back-end.
This commit added the data prefix all the roues for
backend side to fix the issue.