Refactor the token-based structure
Previously, the client-side can get the token
after the login, but missing using token to
authorise users and protect API routes.

This commit add the new middleware called
checkAuth which used to check the token from
the client-side for the first authentication way.

This commit also adds more interaction in the
client-side, for example, if the token be expired,
the user should redirect back to the login page
and login agian for gettting the new token.