Refactor the backend structure to match MVC model
Previously, the structure was not matched the MVC
model which makes all the codes together. The better
way is put the logic codes in the controllers and
only routes in the route files.

This commit creates multiple controllers to match
different fields. The controllers functions that
separate out the code to route requests from the
code that actually processes requests.